// apps/hms-api/src/@generated/prisma-nestjs-graphql/competition-schedule/competition-schedule-min-aggregate.input.ts
import { Field, InputType } from '@nestjs/graphql';

@InputType()
export class CompetitionScheduleMinAggregateInput {
  @Field(() => Boolean, { nullable: true })
  id?: true;

  @Field(() => Boolean, { nullable: true })
  competition_id?: true;

  @Field(() => Boolean, { nullable: true })
  state?: true;

  @Field(() => Boolean, { nullable: true })
  start_at?: true;

  @Field(() => Boolean, { nullable: true })
  end_at?: true;

  @Field(() => Boolean, { nullable: true })
  created_at?: true;

  @Field(() => Boolean, { nullable: true })
  updated_at?: true;

  @Field(() => Boolean, { nullable: true })
  deleted_at?: true;
}